Use tenant token in naming API
Fixes: NSL-632
